Consolidate components tests
