+++ hyperspy/tests/component/test_gaussian.py
@@ -21,9 +21,7 @@ import itertools
 import numpy as np
 import pytest
 
-from hyperspy.components1d import Gaussian
-from hyperspy.signals import Signal1D
-from hyperspy.utils import stack
+import hyperspy.api as hs
 
 sqrt2pi = np.sqrt(2 * np.pi)
 sigma2fwhm = 2 * np.sqrt(2 * np.log(2))
@@ -32,7 +30,7 @@ TRUE_FALSE_2_TUPLE = [p for p in itertools.product((True, False), repeat=2)]
 
 
 def test_function():
-    g = Gaussian()
+    g = hs.model.components1D.Gaussian()
     g.centre.value = 1
     g.sigma.value = 2 / sigma2fwhm
     g.A.value = 3 * sqrt2pi * g.sigma.value
@@ -44,18 +42,18 @@ def test_function():
 @pytest.mark.parametrize(("uniform"), (True, False))
 @pytest.mark.parametrize(("only_current", "binned"), TRUE_FALSE_2_TUPLE)
 def test_estimate_parameters_binned(only_current, binned, lazy, uniform):
-    s = Signal1D(np.empty((100,)))
+    s = hs.signals.Signal1D(np.empty((100,)))
     s.axes_manager.signal_axes[0].is_binned = binned
     axis = s.axes_manager.signal_axes[0]
     axis.scale = 1
     axis.offset = -20
-    g1 = Gaussian(50015.156, 10 / sigma2fwhm, 10)
+    g1 = hs.model.components1D.Gaussian(50015.156, 10 / sigma2fwhm, 10)
     s.data = g1.function(axis.axis)
     if not uniform:
         axis.convert_to_non_uniform_axis()
     if lazy:
         s = s.as_lazy()
-    g2 = Gaussian()
+    g2 = hs.model.components1D.Gaussian()
     if binned and uniform:
         factor = axis.scale
     elif binned:
@@ -72,14 +70,14 @@ def test_estimate_parameters_binned(only_current, binned, lazy, uniform):
 
 
 def test_estimate_parameters_negative_scale():
-    s = Signal1D(np.empty((100,)))
+    s = hs.signals.Signal1D(np.empty((100,)))
     axis = s.axes_manager.signal_axes[0]
     axis.scale = -1
     axis.offset = 100
-    g1 = Gaussian(50015.156, 15 / sigma2fwhm, 50)
+    g1 = hs.model.components1D.Gaussian(50015.156, 15 / sigma2fwhm, 50)
     s.data = g1.function(axis.axis)
 
-    g2 = Gaussian()
+    g2 = hs.model.components1D.Gaussian()
     with pytest.raises(ValueError):
         g2.estimate_parameters(s, 40, 60)
     assert g2.estimate_parameters(s, 90, 10)
@@ -91,17 +89,17 @@ def test_estimate_parameters_negative_scale():
 @pytest.mark.parametrize(("lazy"), (True, False))
 @pytest.mark.parametrize(("binned"), (True, False))
 def test_function_nd(binned, lazy):
-    s = Signal1D(np.empty((100,)))
+    s = hs.signals.Signal1D(np.empty((100,)))
     axis = s.axes_manager.signal_axes[0]
     axis.scale = 1
     axis.offset = -20
-    g1 = Gaussian(50015.156, 10 / sigma2fwhm, 10)
+    g1 = hs.model.components1D.Gaussian(50015.156, 10 / sigma2fwhm, 10)
     s.data = g1.function(axis.axis)
     s.axes_manager.signal_axes[0].is_binned = binned
-    s2 = stack([s] * 2)
+    s2 = hs.stack([s] * 2)
     if lazy:
         s2 = s2.as_lazy()
-    g2 = Gaussian()
+    g2 = hs.model.components1D.Gaussian()
     factor = axis.scale if binned else 1
     g2.estimate_parameters(s2, axis.low_value, axis.high_value, False)
     assert g2._axes_manager[-1].is_binned == binned
@@ -109,38 +107,75 @@ def test_function_nd(binned, lazy):
 
 
 def test_util_fwhm_set():
-    g1 = Gaussian()
+    g1 = hs.model.components1D.Gaussian()
     g1.fwhm = 1.0
     np.testing.assert_allclose(g1.sigma.value, 1.0 / sigma2fwhm)
 
 
 def test_util_fwhm_get():
-    g1 = Gaussian()
+    g1 = hs.model.components1D.Gaussian()
     g1.sigma.value = 1.0
     np.testing.assert_allclose(g1.fwhm, 1.0 * sigma2fwhm)
 
 
 def test_util_fwhm_getset():
-    g1 = Gaussian()
+    g1 = hs.model.components1D.Gaussian()
     g1.fwhm = 1.0
     np.testing.assert_allclose(g1.fwhm, 1.0)
 
 
 def test_util_height_set():
-    g1 = Gaussian()
+    g1 = hs.model.components1D.Gaussian()
     g1.sigma.value = 3.0
     g1.height = 2.0 / sqrt2pi
     np.testing.assert_allclose(g1.A.value, 6)
 
 
 def test_util_height_get():
-    g1 = Gaussian()
+    g1 = hs.model.components1D.Gaussian()
     g1.sigma.value = 4.0
     g1.A.value = sqrt2pi * 8
     np.testing.assert_allclose(g1.height, 2)
 
 
 def test_util_height_getset():
-    g1 = Gaussian()
+    g1 = hs.model.components1D.Gaussian()
     g1.height = 4.0
     np.testing.assert_allclose(g1.height, 4.0)
+
+
+class TestGaussian:
+    def setup_method(self, method):
+        s = hs.signals.Signal1D(np.zeros(1024))
+        s.axes_manager[0].offset = -5
+        s.axes_manager[0].scale = 0.01
+        m = s.create_model()
+        m.append(hs.model.components1D.Gaussian())
+        m[0].sigma.value = 0.5
+        m[0].centre.value = 1
+        m[0].A.value = 2
+        self.m = m
+
+    @pytest.mark.parametrize(("only_current", "binned"), TRUE_FALSE_2_TUPLE)
+    def test_estimate_parameters_binned(self, only_current, binned):
+        self.m.signal.axes_manager[-1].is_binned = binned
+        s = self.m.as_signal()
+        assert s.axes_manager[-1].is_binned == binned
+        g = hs.model.components1D.Gaussian()
+        g.estimate_parameters(s, None, None, only_current=only_current)
+        assert g._axes_manager[-1].is_binned == binned
+        np.testing.assert_allclose(g.sigma.value, 0.5)
+        np.testing.assert_allclose(g.A.value, 2)
+        np.testing.assert_allclose(g.centre.value, 1)
+
+    @pytest.mark.parametrize("binned", (True, False))
+    def test_function_nd(self, binned):
+        self.m.signal.axes_manager[-1].is_binned = binned
+        s = self.m.as_signal()
+        s2 = hs.stack([s] * 2)
+        g = hs.model.components1D.Gaussian()
+        g.estimate_parameters(s2, None, None, only_current=False)
+        assert g._axes_manager[-1].is_binned == binned
+        axis = s.axes_manager.signal_axes[0]
+        factor = axis.scale if binned else 1
+        np.testing.assert_allclose(g.function_nd(axis.axis) * factor, s2.data)
